+++ v2/connection/context.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,8 +17,6 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Jakub Wierzbowski
-//
 
 package connection
 
@@ -33,6 +31,9 @@ const (
 	keyUseQueueTimeout ContextKey = "arangodb-use-queue-timeout"
 	keyMaxQueueTime    ContextKey = "arangodb-max-queue-time-seconds"
 	keyDriverFlags     ContextKey = "arangodb-driver-flags"
+
+	keyAsyncRequest ContextKey = "arangodb-async-request"
+	keyAsyncID      ContextKey = "arangodb-async-id"
 )
 
 // contextOrBackground returns the given context if it is not nil.
@@ -58,4 +59,44 @@ func WithArangoQueueTime(parent context.Context, duration time.Duration) context
 // WithDriverFlags is used to configure additional flags for the `x-arango-driver` header.
 func WithDriverFlags(parent context.Context, value []string) context.Context {
 	return context.WithValue(contextOrBackground(parent), keyDriverFlags, value)
+}
+
+// WithAsync is used to configure a context to make an async operation - requires Connection with Async wrapper!
+func WithAsync(parent context.Context) context.Context {
+	return context.WithValue(contextOrBackground(parent), keyAsyncRequest, true)
+}
+
+// WithAsyncID is used to check an async operation result - requires Connection with Async wrapper!
+func WithAsyncID(parent context.Context, asyncID string) context.Context {
+	return context.WithValue(contextOrBackground(parent), keyAsyncID, asyncID)
+}
+
+//
+// READ METHODS
+//
+
+// IsAsyncRequest returns true if the given context is an async request.
+func IsAsyncRequest(ctx context.Context) bool {
+	if ctx != nil {
+		if v := ctx.Value(keyAsyncRequest); v != nil {
+			if isAsync, ok := v.(bool); ok && isAsync {
+				return true
+			}
+		}
+	}
+
+	return false
+}
+
+// HasAsyncID returns the async Job ID from the given context.
+func HasAsyncID(ctx context.Context) (string, bool) {
+	if ctx != nil {
+		if q := ctx.Value(keyAsyncID); q != nil {
+			if v, ok := q.(string); ok {
+				return v, true
+			}
+		}
+	}
+
+	return "", false
 }